feat(endpoint): support updates
Endpoints may now be updated (patched). It is important to note that
because Go's JSON marshalling can't detect between omitted fields and
fields set to null, they are treated as the same (ommited). Once created
only text fields may be set to NULL by updating the endpoint with an
empty string.

Realm was also removed from the API since there is no clear benefit at
the moment. All use cases so far use the default value for it.