feat: use connection pooling
Using a single connection is not thread-safe and should not be used in
applications like this where several requests may be processed at a
time.

This also makes it more reliable. Previously the db connection could not
recover from a DB server restart. With connection pooling it is now able
to recover. May not be perfect but is good enough atm.

Closes #24.