fix: leaked connections
Creating a transaction and failing to end it with either a commit or
rollback leaves db pool connections in an acquired state, unavailable to
other requests. There were two instances where this was happening, this
resulted in the service eventually becoming unresponsive after the
offending code was called enough times and all the connections were
exhausted.